Procesado de mensajes agregado a vista. Agregado a Cotnrolador y ManejadorMensajes. Todo funca

+++ src/logica/ManejadorDatos.java
@@ -86,7 +86,7 @@ public class ManejadorDatos {
 		            Prestamo prestamo = new Prestamo(
 		            		new Fecha(campos[0]), 
 		            		new Fecha(campos[1]), 
-		            		Integer.parseInt(campos[2]),
+		            		(int)Double.parseDouble(campos[2]),
 		            		campos[8],
 		            		campos[9],
 		            		parseNumero(campos[10]),
